requirements.dev.txt
black==22.1.0
click==8.0.4
memory-profiler==0.60.0
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.1
psutil==5.9.1
tomli==2.0.1
typing_extensions==4.1.1
